chore: extract preset from `rnx-align-deps.js`
